chore: remove eye.pl APIs
