--- eye/pl.js
@@ -1,1 +1,0 @@
-export { default } from './eye.pl'
