+++ package.json
@@ -19,14 +19,12 @@
     "build": "npm run eye:prepare && tsc",
     "prepare": "npm run build",
     "semantic-release": "semantic-release",
-    "eye:fetch": "ts-node scripts/fetch-eye",
     "eye:pvm": "ts-node scripts/generate-pvm",
     "eye:pvm:test": "ts-node scripts/run-pvm",
-    "eye:build:pl": "rollup --config rollup.config.mjs",
-    "eye:build:pvm": "rollup --config rollup.config.pvm.mjs",
-    "eye:prepare": "npm run eye:fetch && npm run eye:build:pl && npm run eye:pvm && npm run eye:build:pvm",
+    "eye:build:pvm": "rollup --config rollup.config.mjs",
+    "eye:prepare": "npm run eye:pvm && npm run eye:build:pvm",
     "eye:update": "ts-node scripts/update",
-    "perf": "ts-node perf/socrates && ts-node perf/socrates-pvm"
+    "perf": "ts-node perf/socrates"
   },
   "repository": {
     "type": "git",
