+++ perf/socrates.ts
@@ -1,24 +1,18 @@
-import { SWIPL, loadEye, queryOnce } from '../dist';
-import { query, data } from '../data/socrates';
+import { data, query } from '../data/socrates';
+import { queryOnce, SwiplEye } from '../dist';
 
-// The results recorded using DELL XPS 15 9520 (32GB RAM) on commit 563f3cfc663b8d746d653e90337002eec74d246e are
-// Initialise SWIPL        : 103.505ms
-// Load and consult EYE    : 464.449ms
-// Load data.n3            : 0.07ms
-// Load query.n3           : 0.021ms
-// Execute query           : 7.207ms
+// The results recorded using DELL XPS 15 9520 (32GB RAM)
+// Initialise SWIPL        : 110.441ms (averages about 50-60ms on subsequent loads in the same session)
+// Load data.n3            : 0.081ms
+// Load query.n3           : 0.03ms
+// Execute query           : 10.882ms
 async function main() {
-  console.log('Testing performance of socrates query using eye.pl')
+  console.log('Testing performance of socrates query using eye.pvm')
 
   // Instantiate a new SWIPL module and log any results it produces to the console
-  console.time('Initialise SWIPL\t');
-  const Module = await SWIPL({ print: () => {}, arguments: ['-q'] });
-  console.timeEnd('Initialise SWIPL\t');
-
-  // Load EYE into the SWIPL Module and run consule("eye.pl").
-  console.time('Load and consult EYE\t');
-  loadEye(Module)
-  console.timeEnd('Load and consult EYE\t');
+  console.time(`Initialise SWIPL with EYE image\t`);
+  const Module = await SwiplEye({ print: () => {} });
+  console.timeEnd(`Initialise SWIPL with EYE image\t`);
 
   // Load the the strings data and query as files data.n3 and query.n3 into the module
   console.time('Load data.n3\t\t');
@@ -32,7 +26,7 @@ async function main() {
   console.time('Execute query\t\t');
   queryOnce(Module, 'main', ['--nope', '--quiet', './data.n3', '--query', './query.n3']);
   console.timeEnd('Execute query\t\t');
-
+  
   console.log()
 }
