--- rollup.config.pvm.mjs
@@ -1,14 +1,0 @@
-import { string } from "rollup-plugin-string";
-
-export default {
-  input: "./eye/pvm",
-  output: {
-    file: "lib/eye.pvm.ts"
-  },
-  plugins: [
-    string({
-      // Required to be specified
-      include: "**/*.pvm",
-    }),
-  ],
-};
