+++ scripts/generate-pvm.ts
@@ -14,16 +14,28 @@ function Uint8ToString(u8a: any) {
   return c.join('');
 }
 
-(async () => {
+async function eyePlString() {
+  const EYE_URL = JSON.parse(fs.readFileSync(path.join(__dirname, '..', 'package.json')).toString()).config.eye.url;
+  const releaseInfo = (await (await fetch(EYE_URL)).json());
+  const res = (await fetch(`https://raw.githubusercontent.com/eyereasoner/eye/${releaseInfo.tag_name}/eye.pl`));
+
+  if (res.status !== 200) {
+    throw new Error(`Error fetching eye.pl: ${await res.text()}`);
+  }
+
+  return res.text();
+}
+
+async function main() {
+  const eye = await eyePlString();
+
   const Module: SWIPLModule = await SWIPL({
     arguments: ['-q', '-f', 'eye.pl'],
-    preRun: (module: SWIPLModule) => {
-      module.FS.writeFile('eye.pl', fs.readFileSync(path.join(__dirname, '..', 'eye', 'eye.pl')));
-    },
+    preRun: (module: SWIPLModule) => module.FS.writeFile('eye.pl', eye),
   });
+
   queryOnce(Module, 'main', ['--image', 'eye.pvm']);
+  fs.writeFileSync(path.join(__dirname, '..', 'eye', 'eye.pvm'), btoa(Uint8ToString(Module.FS.readFile('eye.pvm'))));
+}
 
-  const eyeBuffer = Module.FS.readFile('eye.pvm');
-
-  fs.writeFileSync(path.join(__dirname, '..', 'eye', 'eye.pvm'), btoa(Uint8ToString(eyeBuffer)));
-})();
+main();
